Cleaned up the control scheme system.
