+++ TeamCode/src/main/java/org/nknsd/teamcode/components/drivers/AdvancedWheelDriver.java
@@ -1,11 +1,11 @@
-package org.nknsd.teamcode.drivers;
+package org.nknsd.teamcode.components.drivers;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.nknsd.teamcode.controlSchemes.abstracts.WheelControlScheme;
+import org.nknsd.teamcode.controlSchemes.defaults.WheelControlScheme;
 import org.nknsd.teamcode.frameworks.NKNComponent;
 import org.nknsd.teamcode.components.utility.GamePadHandler;
 import org.nknsd.teamcode.components.sensors.IMUSensor;
