--- TeamCode/src/main/java/org/nknsd/teamcode/controlSchemes/abstracts/WheelControlScheme.java
@@ -1,24 +1,0 @@
-package org.nknsd.teamcode.controlSchemes.abstracts;
-
-import org.nknsd.teamcode.components.utility.GamePadHandler;
-import org.nknsd.teamcode.frameworks.NKNControlScheme;
-
-import java.util.concurrent.Callable;
-
-public abstract class WheelControlScheme extends NKNControlScheme {
-    public abstract Callable<Boolean> gearUp();
-    public abstract Callable<Boolean> gearDown();
-    public abstract Callable<Boolean> resetAngle();
-    public Callable<Boolean> specFancyDeposit(){
-        return () -> false;
-    };
-
-    public Callable<Boolean> initDisableAutoFix() {
-        return new Callable<Boolean>() {
-            @Override
-            public Boolean call() throws Exception {
-                return GamePadHandler.GamepadButtons.Y.detect(gamePadHandler.getGamePad1());
-            }
-        };
-    }
-}
