+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/AbsoluteMovementTestProgram.java
@@ -3,13 +3,13 @@ package org.nknsd.teamcode.programs.tests;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.nknsd.teamcode.controlSchemes.reals.DefaultWheelController;
+import org.nknsd.teamcode.controlSchemes.defaults.WheelControlScheme;
 import org.nknsd.teamcode.frameworks.NKNComponent;
 import org.nknsd.teamcode.components.sensors.FlowSensor;
 import org.nknsd.teamcode.components.utility.GamePadHandler;
 import org.nknsd.teamcode.components.sensors.IMUSensor;
 import org.nknsd.teamcode.components.handlers.WheelHandler;
-import org.nknsd.teamcode.drivers.AdvancedWheelDriver;
+import org.nknsd.teamcode.components.drivers.AdvancedWheelDriver;
 import org.nknsd.teamcode.frameworks.NKNProgram;
 
 import java.util.List;
@@ -38,7 +38,7 @@ public class AbsoluteMovementTestProgram extends NKNProgram {
         components.add(imuSensor);
         //telemetryEnabled.add(imuComponent);
 
-        DefaultWheelController wheelController = new DefaultWheelController();
+        WheelControlScheme wheelController = new WheelControlScheme();
         wheelController.link(gamePadHandler);
 
         // Wheel Driver
